work around windows issues
git-svn-id: https://svn.apache.org/repos/asf/maven/shared/trunk@635721 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-verifier/src/main/java/org/apache/maven/it/util/ResourceExtractor.java
@@ -4,10 +4,15 @@
  */
 package org.apache.maven.it.util;
 
-import java.io.*;
-import java.net.*;
-import java.util.*;
-import java.util.zip.*;
+import java.io.File;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.util.Enumeration;
+import java.util.zip.ZipEntry;
+import java.util.zip.ZipFile;
 
 
 /* @todo this can be replaced with plexus-archiver */
@@ -15,9 +20,24 @@ public class ResourceExtractor {
     
     public static File simpleExtractResources(Class cl, String resourcePath) throws IOException {
         String tempDirPath = System.getProperty( "maven.test.tmpdir", System.getProperty( "java.io.tmpdir" ) );
-        File tempDir = new File(tempDirPath);
+        File tempDir = null;
+        if (Os.isFamily( "windows" ))
+        {
+           //there are issues deleting on windows. use a random path.
+            tempDir = new File(tempDirPath,""+ Math.round( Math.random()*100000000 ));
+        }
+        else
+        {
+            tempDir = new File(tempDirPath);
+        }
+        
         File testDir = new File( tempDir, resourcePath );
-        FileUtils.deleteDirectory( testDir );
+
+        if (!Os.isFamily( "windows" ))
+        {
+            FileUtils.deleteDirectory( testDir );
+        }
+        
         testDir = ResourceExtractor.extractResourcePath(cl, resourcePath, tempDir, false);
         return testDir;
     }
